Create report conductor with police button
Needs repairing the calculation of chances in client (danger-indicator.component).

+++ server/src/main/migrations/20180118000000-create-edge.ts
@@ -12,7 +12,7 @@ export = new NewTableMigration("Edges", (types: DataTypes): DefineAttributes => 
          model: "Stations",
          key: "id",
       },
-      type: types.INTEGER,
+      type: types.BIGINT,
    },
    toStationId: {
       allowNull: false,
@@ -20,7 +20,7 @@ export = new NewTableMigration("Edges", (types: DataTypes): DefineAttributes => 
          model: "Stations",
          key: "id",
       },
-      type: types.INTEGER,
+      type: types.BIGINT,
    },
    travelTimeMs: {
       allowNull: false,
